refactor : extract social icons from common icons
