+++ src/components/follow/Icons.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import Icon from '@ant-design/icons';
+import { PlusSvg, ChatSmallSvg } from './Svgs';
+
+/*eslint-disable */
+
+export const PlusIcon = (props) => <Icon component={PlusSvg} {...props} />;
+
+export const ChatSmallIcon = (props) => (
+  <Icon component={ChatSmallSvg} {...props} />
+);
