+++ src/components/follow/Social.jsx
@@ -1,26 +1,20 @@
 import React from 'react';
 import { useRecoilValue } from 'recoil';
-import { Tabs, Input } from 'antd';
+import { Tabs } from 'antd';
 import SocialList from './SocialList';
+import SocialSearch from './SocialSearch';
 
 import { followersState, followingsState } from '../../state/follow';
 
 const { TabPane } = Tabs;
-const { Search } = Input;
 
 export default function Social() {
   const followers = useRecoilValue(followersState);
   const followings = useRecoilValue(followingsState);
-  const onSearch = (value) => console.log(value);
 
   return (
     <>
-      <Search
-        placeholder="input search text"
-        allowClear
-        onSearch={onSearch}
-        style={{ marginBottom: '1rem' }}
-      />
+      <SocialSearch />
       {followers && followings && (
         <Tabs>
           <TabPane tab={`팔로워(${followers.length}명)`} key="1">
