+++ src/components/follow/Svgs.jsx
@@ -1,0 +1,53 @@
+import React from 'react';
+
+export const PlusSvg = () => (
+  <svg width="25" height="25" viewBox="0 0 25 25" fill="none">
+    <rect
+      x="0.855"
+      y="0.855"
+      width="23.29"
+      height="23.29"
+      rx="11.645"
+      stroke="white"
+      strokeWidth="1.71"
+    />
+    <rect
+      x="11.2129"
+      y="6.05273"
+      width="2.57891"
+      height="12.8946"
+      rx="1.28946"
+      fill="white"
+    />
+    <rect
+      x="18.9482"
+      y="11.2109"
+      width="2.57891"
+      height="12.8946"
+      rx="1.28946"
+      transform="rotate(90 18.9482 11.2109)"
+      fill="white"
+    />
+  </svg>
+);
+
+export const ChatSmallSvg = () => (
+  <svg width="31" height="22" viewBox="0 0 31 22" fill="none">
+    <path
+      d="M19.9286 0H11.0714C4.95446 0 0 4.92196 0 10.9988C0 17.0756 4.95446 21.9976 11.0714 21.9976H29.3393C30.2582 21.9976 31 21.2607 31 20.3478V10.9988C31 4.92196 26.0455 0 19.9286 0ZM1.66071 10.9988C1.66071 5.84585 5.88446 1.64982 11.0714 1.64982H19.9286C25.1155 1.64982 29.3393 5.84585 29.3393 10.9988V17.5101L26.4829 20.3478H11.0714C5.88446 20.3478 1.66071 16.1517 1.66071 10.9988Z"
+      fill="white"
+    />
+    <path
+      d="M8.28767 12.6492C9.20485 12.6492 9.94838 11.9106 9.94838 10.9994C9.94838 10.0883 9.20485 9.34961 8.28767 9.34961C7.37048 9.34961 6.62695 10.0883 6.62695 10.9994C6.62695 11.9106 7.37048 12.6492 8.28767 12.6492Z"
+      fill="white"
+    />
+    <path
+      d="M15.5006 12.6492C16.4177 12.6492 17.1613 11.9106 17.1613 10.9994C17.1613 10.0883 16.4177 9.34961 15.5006 9.34961C14.5834 9.34961 13.8398 10.0883 13.8398 10.9994C13.8398 11.9106 14.5834 12.6492 15.5006 12.6492Z"
+      fill="white"
+    />
+    <path
+      d="M22.69 12.6492C23.6072 12.6492 24.3507 11.9106 24.3507 10.9994C24.3507 10.0883 23.6072 9.34961 22.69 9.34961C21.7728 9.34961 21.0293 10.0883 21.0293 10.9994C21.0293 11.9106 21.7728 12.6492 22.69 12.6492Z"
+      fill="white"
+    />
+  </svg>
+);
